+++ packages/docs/.storybook/preview.js
@@ -10,5 +10,17 @@ export const parameters = {
   },
   docs: {
     theme: themes.dark
-  }
+  }, backgrounds: {
+    default: 'dark',
+    values: [
+      {
+        name: 'dark',
+        value: '#1E1E1E',
+      },
+      {
+        name: 'light',
+        value: '#FFF',
+      },
+    ],
+  },
 }