+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @omn-ui/react
 
+## 3.0.0
+
+### Major Changes
+
+- Toast and Tooltip component added
+
 ## 2.0.0
 
 ### Major Changes