+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @omn-ui/tokens
 
+## 3.0.0
+
+### Major Changes
+
+- Toast and Tooltip component added
+
 ## 2.1.0
 
 ### Minor Changes